chore(backend): move authenticate controller to controllers folder
