+++ backend/src/modules/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
 import { DatabaseModule } from '~/database.module';
 import { AuthenticateAccountService } from './services/authenticate-account/authenticate-account.service';
-import { AuthenticateAccountController } from './authenticate-account/authenticate-account.controller';
+import { AuthenticateAccountController } from './controllers/authenticate-account/authenticate-account.controller';
 import { HashProvider } from '~/common/providers/hash.provider';
 
 @Module({
